Create retrieves_info_from_a_thumbcacheDB.ps1
This script retrieves the filename and last access time of all thumbnails in a specified thumbcache database file, using the SQLite module in PowerShell to query the database. To use this script, you would need to replace the thumbcachePath variable with the path to the thumbcache database file you want to analyze.